Update DDoS_Attack.md
This is a breakdown of the Diagram:

The Attacker is the one initiating the Cyberattack, specifically DDoS, which is a denial of service attack.

The Method they will be utilizing for this attack is a BotNet. A BotNet is a tool for carrying out DDoS attacks. The BotNet requires some information from the Attacker in order to carry out the attack. The attacker will need to provide the length of time for the attack to last, either an IP address or web address, and the attack type.

From this point, the BotNet will attack the selected IP address or web address by flooding those targets with a lot of traffic. So much traffic to where it causes the Web server to bottleneck and effectively render its service unusable, hence the denial of service.

There are ways to not only detect but also act against said attacks. The webpage has a firewall which all traffic is filtered too. I view the firewall as a bouncer saying who is allowed in and who is denied access. The firewall has set parameters, and if a certain IP address fails to meet that it will block it. The firewall plays a key defensive role in protecting against DDoS. The firewall can inspect traffic, monitor traffic levels, and filter based on geographical location.